requirements.txt
Flask==2.3.2
psycopg2-binary==2.9.6
azure-storage-blob==12.15.0
gunicorn==20.1.0
